requirements.txt
fastapi
uvicorn
openai
pinecone
requests
python-dotenv
fastmcp
pydantic
tenacity>=8.2.3 